Move generic edit log components to common app, reorder URL structure, rename url parameter to make explicit
