+++ bcap/src/bcap/components/pages/details/ArchaeologicalSite/sections/DetailsSection4.vue
@@ -6,7 +6,7 @@ import { getDisplayValue, isAliasedNodeData, isEmpty } from '@/bcap/util.ts';
 import {
     useTileEditLog,
     useSingleTileEditLog,
-} from '@/bcap/composables/useTileEditLog.ts';
+} from '@/bcgov_arches_common/composables/useTileEditLog.ts';
 import type { EditLogData } from '@/bcgov_arches_common/types.ts';
 import { EDIT_LOG_FIELDS } from '@/bcgov_arches_common/constants.ts';
 import StandardDataTable from '@/bcgov_arches_common/components/StandardDataTable/StandardDataTable.vue';
