+++ bcap/src/bcap/components/pages/details/ArchaeologicalSite/sections/DetailsSection6.vue
@@ -3,7 +3,7 @@ import { computed, toRef } from 'vue';
 import DetailsSection from '@/bcap/components/DetailsSection/DetailsSection.vue';
 import EmptyState from '@/bcap/components/EmptyState.vue';
 import { useHierarchicalData } from '@/bcap/composables/useHierarchicalData.ts';
-import { useTileEditLog } from '@/bcap/composables/useTileEditLog.ts';
+import { useTileEditLog } from '@/bcgov_arches_common/composables/useTileEditLog.ts';
 import type { EditLogData } from '@/bcgov_arches_common/types.ts';
 import { EDIT_LOG_FIELDS } from '@/bcgov_arches_common/constants.ts';
 import StandardDataTable from '@/bcgov_arches_common/components/StandardDataTable/StandardDataTable.vue';
