+++ bcap/src/bcap/components/pages/details/SiteVisit/sections/SiteVisitDetailsSection2.vue
@@ -7,7 +7,7 @@ import StandardDataTable from '@/bcgov_arches_common/components/StandardDataTabl
 import {
     useTileEditLog,
     useSingleTileEditLog,
-} from '@/bcap/composables/useTileEditLog.ts';
+} from '@/bcgov_arches_common/composables/useTileEditLog.ts';
 import type { EditLogData } from '@/bcgov_arches_common/types.ts';
 import { EDIT_LOG_FIELDS } from '@/bcgov_arches_common/constants.ts';
 import type { SiteVisitSchema } from '@/bcap/schema/SiteVisitSchema.ts';
