+++ bcap/src/bcap/components/pages/details/SiteVisit/sections/SiteVisitDetailsSection3.vue
@@ -7,7 +7,7 @@ import { getDisplayValue, isEmpty } from '@/bcap/util.ts';
 import {
     useTileEditLog,
     useSingleTileEditLog,
-} from '@/bcap/composables/useTileEditLog.ts';
+} from '@/bcgov_arches_common/composables/useTileEditLog.ts';
 import type { EditLogData } from '@/bcgov_arches_common/types.ts';
 import { EDIT_LOG_FIELDS } from '@/bcgov_arches_common/constants.ts';
 import type { SiteVisitSchema } from '@/bcap/schema/SiteVisitSchema.ts';
